refactor: Move SyncSettingsViewModel to it's own file
